requirements.txt
dotenv
python-telegram-bot==20.7
python-dotenv==1.0.0
sqlalchemy==2.0.23